+++ History.md
@@ -1,6 +1,12 @@
 0.2.6 / 2012-05-23
 ==================
 
+* Fix: `addRule(String)` now properly validates the remaining size
+* Removed first rule validation enforcement
+
+0.2.6 / 2012-05-23
+==================
+
 * Fix: `addRule(null)` throws an error
 * Fix: `addRule([1,2])` returns proper token
 * Fix: first rule validation enforced (waits for more data if required, which means rules starting with an array of numbers is equivalent to a rule with the max of those numbers: addRule([1,2]) <=> addRule(2))